Move files to algorithm and tests instead of package moral
