--- networkx/algorithms/moral/__init__.py
@@ -1,19 +1,0 @@
-# __init__.py - package containing moralization
-#
-# Copyright 2016-2018 NetworkX developers.
-#
-# This file is part of NetworkX.
-#
-# NetworkX is distributed under a BSD license; see LICENSE.txt for more
-# information.
-"""Moralization of a directed graph.
-
-    .. warning:: The moral submodule is not imported in the top-level
-        ``networkx``.
-
-    These functions can be imported with
-        ``from networkx.algorithms import moral.
-
-"""
-
-from networkx.algorithms.moral.moralize import *
